@@ -20,10 +20,10 @@ class MyStyles {
       AppBarTheme(
         elevation: 0,
         titleTextStyle:
-            getTextTheme(isLightTheme: isLightTheme).bodyText1!.copyWith(
-                  color: Colors.white,
-                  fontSize: MyFonts.appBarTittleSize,
-                ),
+        getTextTheme(isLightTheme: isLightTheme).bodyMedium!.copyWith(
+          color: Colors.white,
+          fontSize: MyFonts.appBarTittleSize,
+        ),
         iconTheme: IconThemeData(
             color: isLightTheme
                 ? LightThemeColors.appBarIconsColor
@@ -35,61 +35,48 @@ class MyStyles {
 
   ///text theme
   static TextTheme getTextTheme({required bool isLightTheme}) => TextTheme(
-        button: MyFonts.buttonTextStyle
-            .copyWith(fontSize: MyFonts.buttonTextSize),
-        bodyText1: (MyFonts.bodyTextStyle).copyWith(
-            fontWeight: FontWeight.bold,
-            fontSize: MyFonts.body1TextSize,
-            color: isLightTheme
-                ? LightThemeColors.bodyTextColor
-                : DarkThemeColors.bodyTextColor),
-        bodyText2: (MyFonts.bodyTextStyle).copyWith(
-            fontSize: MyFonts.body2TextSize,
-            color: isLightTheme
-                ? LightThemeColors.bodyTextColor
-                : DarkThemeColors.bodyTextColor),
-        headline1: (MyFonts.headlineTextStyle).copyWith(
-            fontSize: MyFonts.headline1TextSize,
-            fontWeight: FontWeight.bold,
-            color: isLightTheme
-                ? LightThemeColors.headlinesTextColor
-                : DarkThemeColors.headlinesTextColor),
-        headline2: (MyFonts.headlineTextStyle).copyWith(
-            fontSize: MyFonts.headline2TextSize,
-            fontWeight: FontWeight.bold,
-            color: isLightTheme
-                ? LightThemeColors.headlinesTextColor
-                : DarkThemeColors.headlinesTextColor),
-        headline3: (MyFonts.headlineTextStyle).copyWith(
-            fontSize: MyFonts.headline3TextSize,
-            fontWeight: FontWeight.bold,
-            color: isLightTheme
-                ? LightThemeColors.headlinesTextColor
-                : DarkThemeColors.headlinesTextColor),
-        headline4: (MyFonts.headlineTextStyle).copyWith(
-            fontSize: MyFonts.headline4TextSize,
-            fontWeight: FontWeight.bold,
-            color: isLightTheme
-                ? LightThemeColors.headlinesTextColor
-                : DarkThemeColors.headlinesTextColor),
-        headline5: (MyFonts.headlineTextStyle).copyWith(
-            fontSize: MyFonts.headline5TextSize,
-            fontWeight: FontWeight.bold,
-            color: isLightTheme
-                ? LightThemeColors.headlinesTextColor
-                : DarkThemeColors.headlinesTextColor),
-        headline6: (MyFonts.headlineTextStyle).copyWith(
-            fontSize: MyFonts.headline6TextSize,
-            fontWeight: FontWeight.bold,
-            color: isLightTheme
-                ? LightThemeColors.headlinesTextColor
-                : DarkThemeColors.headlinesTextColor),
-        caption: TextStyle(
-            color: isLightTheme
-                ? LightThemeColors.captionTextColor
-                : DarkThemeColors.captionTextColor,
-            fontSize: MyFonts.captionTextSize),
-      );
+    labelLarge: MyFonts.buttonTextStyle.copyWith(
+      fontSize: MyFonts.buttonTextSize,
+    ),
+    bodyLarge: (MyFonts.bodyTextStyle).copyWith(
+      fontWeight: FontWeight.bold,
+      fontSize: MyFonts.bodyLargeSize,
+      color: isLightTheme
+          ? LightThemeColors.bodyTextColor
+          : DarkThemeColors.bodyTextColor,
+    ),
+    bodyMedium: (MyFonts.bodyTextStyle).copyWith(
+      fontSize: MyFonts.bodyMediumSize,
+      color: isLightTheme
+          ? LightThemeColors.bodyTextColor
+          : DarkThemeColors.bodyTextColor,
+    ),
+    displayLarge: (MyFonts.displayTextStyle).copyWith(
+      fontSize: MyFonts.displayLargeSize,
+      fontWeight: FontWeight.bold,
+      color: isLightTheme
+          ? LightThemeColors.displayTextColor
+          : DarkThemeColors.displayTextColor,
+    ),
+    bodySmall: TextStyle(
+        color: isLightTheme
+            ? LightThemeColors.bodySmallTextColor
+            : DarkThemeColors.bodySmallTextColor,
+        fontSize: MyFonts.bodySmallTextSize),
+    displayMedium: (MyFonts.displayTextStyle).copyWith(
+        fontSize: MyFonts.displayMediumSize,
+        fontWeight: FontWeight.bold,
+        color: isLightTheme
+            ? LightThemeColors.displayTextColor
+            : DarkThemeColors.displayTextColor),
+    displaySmall: (MyFonts.displayTextStyle).copyWith(
+      fontSize: MyFonts.displaySmallSize,
+      fontWeight: FontWeight.bold,
+      color: isLightTheme
+          ? LightThemeColors.displayTextColor
+          : DarkThemeColors.displayTextColor,
+    ),
+  );
 
   static ChipThemeData getChipTheme({required bool isLightTheme}) {
     return ChipThemeData(
