@@ -27,9 +27,9 @@ class DarkThemeColors
 
   //TEXT
   static const Color bodyTextColor = Color(0xFF96A7AF);
-  static const Color headlinesTextColor = Colors.white;
-  static const Color captionTextColor = Colors.black;
-  static const Color hintTextColor = Color(0xff686868);
+  static const Color displayTextColor = Colors.white;
+  static const Color bodySmallTextColor = Color(0xFF96A7AF);
+  static const Color hintTextColor =  Color(0xff686868);
 
   //chip
   static const Color chipBackground = primaryColor;
