@@ -1,14 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../app/data/local/my_shared_pref.dart';
-import '../translations/localization_service.dart';
-
 // todo configure text family and size
 class MyFonts
 {
   // return the right font depending on app language
-  static TextStyle get getAppFontType => LocalizationService.supportedLanguagesFontsFamilies[MySharedPref.getCurrentLocal().languageCode]!;
+  static TextStyle get getAppFontType => const TextStyle(fontFamily: 'Poppins');
 
   // headlines text font
   static TextStyle get headlineTextStyle => getAppFontType;
@@ -28,17 +25,21 @@ class MyFonts
   // appbar font size
   static double get appBarTittleSize => 18.sp;
 
+  // headlines text font
+  static TextStyle get displayTextStyle => getAppFontType;
+
+  // body font size
+  static double get bodySmallTextSize => 12.sp;
+  static double get bodyMediumSize => 14.sp; // default font
+  static double get bodyLargeSize => 20.sp;
+  // display font size
+  static double get displayLargeSize => 24.sp;
+  static double get displayMediumSize => 18.sp;
+  static double get displaySmallSize => 14.sp;
+
   // body font size
   static double get body1TextSize => 20.sp;
   static double get body2TextSize => 14.sp;
-
-  // headlines font size
-  static double get headline1TextSize => 50.sp;
-  static double get headline2TextSize => 42.sp;
-  static double get headline3TextSize => 24.sp;
-  static double get headline4TextSize => 18.sp;
-  static double get headline5TextSize => 16.sp;
-  static double get headline6TextSize => 14.sp;
 
   //button font size
   static double get buttonTextSize => 16.sp;
